split types file, add more logger messages
