+++ docs/assets/style.css
@@ -492,8 +492,7 @@
     }
     body {
         background: var(--color-background);
-        font-family:
-            -apple-system, BlinkMacSystemFont, "Segoe UI", "Noto Sans",
+        font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Noto Sans",
             Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji";
         font-size: 16px;
         color: var(--color-text);
