+++ src/output.css
@@ -156,8 +156,7 @@ html,
     -o-tab-size: 4;
     tab-size: 4;
     /* 3 */
-    font-family:
-        ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
+    font-family: ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
         "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
     /* 4 */
     font-feature-settings: normal;
@@ -247,8 +246,7 @@ code,
 kbd,
 samp,
 pre {
-    font-family:
-        ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas,
+    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas,
         "Liberation Mono", "Courier New", monospace;
     /* 1 */
     font-feature-settings: normal;
