+++ src/types/types.ts
@@ -1,61 +1,4 @@
-/**
- * interface DicomTag
- * @property tagName - Name of the DICOM tag
- * @property value - Value of the DICOM tag
- * @returns DicomTag
- */
-export interface DicomTag {
-    tagName: string;
-    value: string | DicomTag[];
-    hidden?: boolean;
-}
-
-/**
- * interface DicomTableProps
- * @property dicomData - DICOM data
- * @returns DicomTableProps
- * @description - Interface for the DicomTable component
- */
-export interface DicomTableProps {
-    dicomData: { [key: string]: DicomTag };
-    fileName: string;
-    updateTableData: (data: any) => void;
-    newTableData: any[];
-    clearData: () => void;
-}
-
-/**
- * interface DicomTableRowProps
- * @property row - DICOM tag row
- * @property index - Index of the row
- * @property onUpdateValue - Function to update the value of the row
- * @property nested - Boolean to check if the row is nested
- * @property updated - Boolean to check if the row has been updated
- * @property level - Nesting level of the row
- */
-export interface DicomTableRowProps {
-    row: {
-        tagId: string;
-        tagName: string;
-        value: string | { [tags: string]: DicomTag };
-    };
-    index: number;
-    onUpdateValue: (tagId: string, newValue: string) => void;
-    nested?: boolean;
-    updated?: boolean;
-    level?: number;
-}
-
-/**
- * interface DicomTableRowProps
- * @property row - DICOM tag row
- * @property index - Index of the row
- * @returns DicomTableRowProps
- * @description - Interface for the DicomTableRow component
- */
-export interface CustomFile {
-    name: string;
-}
+import { CustomFile } from "./FileTypes";
 
 /**
  * interface SidebarProps
@@ -76,35 +19,6 @@ export interface SidebarProps {
 }
 
 /**
- * interface FileListProps
- * @property files - Array of files
- * @property currentFileIndex - Index of the currently viewed file
- * @property onFileSelect - Function to handle file selection
- * @property series - Boolean to check if the files are being edited as a series
- */
-export interface FileListProps {
-    files: CustomFile[];
-    currentFileIndex: number;
-    onFileSelect: (index: number) => void;
-    openModal: (value: boolean) => void;
-    series: boolean;
-}
-
-/**
- * interface FileTableProps
- * @property files - Array of files
- * @property currentFileIndex - Index of the currently viewed file
- * @property onFileSelect - Function to handle file selection
- * @property series - Boolean to check if the files are being edited as a series
- */
-export interface FileTableProps {
-    files: any[];
-    currentFileIndex: number;
-    onFileSelect: (index: number) => void;
-    series: boolean;
-}
-
-/**
  * interface AppState
  * @property files - Array of files
  * @property dicomData - Array of dicom data
@@ -115,31 +29,6 @@ export interface FileTableProps {
 export interface AppState {
     files: CustomFile[];
     dicomData: any[];
-    currentFileIndex: number;
-}
-
-/**
- * interface FileNavigationProps
- */
-export interface FileNavigationProps {
-    currentFileIndex: number;
-    fileCount: number;
-    onPrevFile: () => void;
-    onNextFile: () => void;
-}
-
-/**
- * interface FileUploaderProps
- */
-export interface FileUploaderProps {
-    onFileUpload: (files: CustomFile[], dicomData: any[]) => void;
-}
-
-/**
- * interface FileHeaderProps
- */
-export interface FileHeaderProps {
-    files: CustomFile[];
     currentFileIndex: number;
 }
